fix: patrocinio now works on both screens
